tests/submission/W3C/domhtml/HTMLParamElement01.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - PARAM</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/the-param-element.html#dom-param-name" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<p>
<object>
<param name="image3" type="image/gif" value="image/file.gif" valuetype="ref">
</object>
</p>


<script>
var domtest = async_test("The  name attribute specifies the name of the run-time parameter.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vname;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("param");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      vname = testNode.name;

      assert_equals(vname,"image3");
       
    domtest.done();
}));

</script>
<p>The  name attribute specifies the name of the run-time parameter.

    Retrieve the name attribute and examine its value.</p>
<div id="log"></div>
</body></html>
